make the seed 64 bit

+++ solver/rng.h
@@ -5,7 +5,7 @@ struct rng {
         unsigned int i;
 };
 
-void rng_init(struct rng *rng, int seed);
+void rng_init(struct rng *rng, uint64_t seed);
 uint32_t rng_rand_u32(struct rng *rng);
 
 int rng_rand(struct rng *rng, int min, int max);
